backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
email-validator
resend
python-jose[cryptography]
passlib[bcrypt]
python-multipart
